requirements.txt
# ExoHunt Backend Requirements

# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# ML/AI
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Astronomy packages
lightkurve>=2.4.0
astropy>=5.3.0
astroquery>=0.4.6

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
scipy>=1.10.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL (optional)

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
Pillow>=10.0.0

# Development
ipython>=8.12.0
jupyter>=1.0.0
black>=23.3.0
flake8>=6.0.0

# Optional: For production
gunicorn>=20.1.0
whitenoise>=6.4.0
celery>=5.3.0
redis>=4.5.0
